Refactored InMemory Test KEK manager
