+++ kroxylicious-kms-provider-hashicorp-vault-test-support/src/main/java/io/kroxylicious/kms/provider/hashicorp/vault/model/CreatePolicyRequest.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.hashicorp.vault;
+package io.kroxylicious.kms.provider.hashicorp.vault.model;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -12,9 +12,9 @@ import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Objects;
 
-record CreatePolicyRequest(String policy) {
+public record CreatePolicyRequest(String policy) {
 
-    static CreatePolicyRequest fromInputStream(InputStream is) {
+    public static CreatePolicyRequest fromInputStream(InputStream is) {
         Objects.requireNonNull(is);
         try {
             return new CreatePolicyRequest(new String(is.readAllBytes(), StandardCharsets.UTF_8));
