+++ kroxylicious-kms-provider-hashicorp-vault-test-support/src/main/java/io/kroxylicious/kms/provider/hashicorp/vault/model/EnableEngineRequest.java
@@ -4,6 +4,6 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.hashicorp.vault;
+package io.kroxylicious.kms.provider.hashicorp.vault.model;
 
-record EnableEngineRequest(String type) {}
+public record EnableEngineRequest(String type) {}
