+++ kroxylicious-kms-provider-hashicorp-vault-test-support/src/main/java/io/kroxylicious/kms/provider/hashicorp/vault/model/UpdateKeyConfigRequest.java
@@ -4,8 +4,8 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.hashicorp.vault;
+package io.kroxylicious.kms.provider.hashicorp.vault.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-record UpdateKeyConfigRequest(@JsonProperty("deletion_allowed") boolean deletionAllowed) {}
+public record UpdateKeyConfigRequest(@JsonProperty("deletion_allowed") boolean deletionAllowed) {}
