+++ kroxylicious-systemtests/src/main/java/io/kroxylicious/systemtests/resources/kms/vault/KubeVaultTestKmsFacade.java
@@ -18,6 +18,8 @@ import io.kroxylicious.systemtests.utils.VersionComparator;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 
+import static io.kroxylicious.kms.provider.hashicorp.vault.VaultKmsTestUtils.VAULT_ROOT_TOKEN;
+
 /**
  * KMS Facade for Vault running inside Kube.
  */
